API reference

MetaDataGraphs.AbstractDataGraph — Type

AbstractDataGraph{T,VL,VD,ED,GD} <: AbstractGraph{T}

General template for graphs with metadata.

T<:Integer is the type of vertices

VL is the type of vertex labels (cannot be a subtype of Integer)

VD is the type of vertex data objects

ED is the type of edge data objects

GD is the type of the graph data object

MetaDataGraphs.DataDiGraph — Type

DataDiGraph{T,VL,VD,ED,GD} <: AbstractDataGraph{T}

Structure for graphs with metadata based on adjacency list storage.

Fields

ne::Int: number of edges

fadjlist::Vector{Vector{Int}}: forward adjacency lists

badjlist::Vector{Vector{Int}}: backward adjacency lists

labels::Vector{VL}: list of vertex labels

vertices::Dict{VL,T}: dictionary mapping each vertex label to the associated integer index v

vertex_data::Vector{VD}: list of vertex data objects, indexed by integers v

edge_data::Vector{Vector{ED}}: list of edge data objects, indexed by s first and d_index second, where d_index is the rank of d among the outneighbors of s

graph_data::GD: single graph data object

MetaDataGraphs.DataDiGraph — Method

DataDiGraph{T,VL,VD,ED}(graph_data)

Constructor taking only label and data types to create an empty DataDiGraph.

Base.getindex — Method

getindex(g, label)

Alias for get_vertex(g, label), intended for dictionary-like use.

Base.haskey — Method

haskey(g, label)

Check whether a vertex with label label exists.

Graphs.SimpleGraphs.add_edge! — Method

add_edge!(g, label_s, label_d, data)

Add an edge between the vertices with labels label_s and label_d, along with the associated data.

Graphs.SimpleGraphs.add_edge! — Method

add_edge!(g::DataDiGraph, s, d, data)

Add an edge between vertices s and d, along with the associated data.

Graphs.SimpleGraphs.add_vertex! — Method

add_vertex!(g::DataDiGraph, label, data)

Add a vertex with label label, along with the associated data.

Graphs.SimpleGraphs.rem_edge! — Method

rem_edge!(g, label_s, label_d)

Remove the edge between the vertices with labels label_s and label_d if it exists.

Graphs.SimpleGraphs.rem_vertex! — Method

rem_vertex!(g, label)

Remove the vertex with label label if it exists.

Graphs.has_edge — Method

has_edge(g, label_s, label_d)

Check whether an edge exists between the vertices with labels label_s and label_d.

Graphs.has_vertex — Method

has_vertex(g, label)

Check whether a vertex with label label exists by calling haskey(g, label).

Graphs.weights — Method

weights(g)

Compute the edge weights for an AbstractDataGraph.

If the edge data is of type ED = Nothing, return a lazy matrix full of ones.

If the edge data is of type ED <: Real, fill a sparse matrix with it.

If the edge data is of another type ED, try to use the attribute edge_data.weight for each edge.

MetaDataGraphs.get_data — Method

get_data(g, s, d)

Retrieve edge-level metadata for the edge between the vertices s and d.

MetaDataGraphs.get_data — Method

get_data(g, v)

Retrieve vertex-level metadata for the vertex v.

MetaDataGraphs.get_data — Method

get_data(g)

Retrieve graph-level metadata.

MetaDataGraphs.get_data — Method

get_data(g, label_s, label_d)

Retrieve edge-level metadata for the edge between the vertices with labels label_s and label_d.

MetaDataGraphs.get_data — Method

get_data(g, label)

Retrieve vertex-level metadata for the vertex with label label.

MetaDataGraphs.get_label — Method

get_label(g, v)

Retrieve the label associated with vertex v.

MetaDataGraphs.get_vertex — Method

get_vertex(g, label)

Retrieve the vertex associated with label `label.

MetaDataGraphs.set_data! — Method

set_data!(g, s, d, data)

Set edge-level metadata to value data for edge (s, d).

MetaDataGraphs.set_data! — Method

set_data!(g, label_s, label_d, data)

Set edge-level metadata to value data for the edge between the vertices with labels label_s and label_d.

MetaDataGraphs.set_data! — Method

set_data!(g, data)

Set graph-level metadata to value data.

MetaDataGraphs.set_data! — Method

set_data!(g, v, data)

Set vertex-level metadata to value data for the vertex v.

MetaDataGraphs.set_data! — Method

set_data!(g, label, data)

Set vertex-level metadata to value data for the vertex with label label.