AbstractDataGraph

Type

MetaDataGraphs.AbstractDataGraph — Type

AbstractDataGraph{T,VL,VD,ED,GD} <: AbstractGraph{T}

General template for graphs with metadata.

T<:Integer is the type of vertices

VL is the type of vertex labels (cannot be a subtype of Integer)

VD is the type of vertex data objects

ED is the type of edge data objects

GD is the type of the graph data object

Graphs.jl interface

Graphs.SimpleGraphs.add_edge! — Method

add_edge!(g, s, d)

Shortcut for add_edge!(g, s, d, nothing), to use in graphs where edges have no metadata.

Graphs.SimpleGraphs.add_edge! — Method

add_edge!(g::DataDiGraph, label_s, label_d, data)

Add an edge to g by specifying the labels label_s and label_d of both vertices, along with the associated data.

Graphs.SimpleGraphs.add_edge! — Method

add_edge!(g, label_s, label_d)

Shortcut for add_edge!(g, label_s, label_d, nothing), to use in graphs where edges have no metadata.

Graphs.SimpleGraphs.add_vertex! — Method

add_vertex!(g, label)

Shortcut for add_vertex(g, label, nothing), to use in graphs where vertices have no metadata.

Graphs.has_edge — Method

has_edge(g, label_s, label_d)

Check whether an edge exists between the vertices with labels label_s and label_d.

Graphs.has_vertex — Method

has_vertex(g, label)

Check whether a vertex with label label exists by calling haskey(g, label).

Metadata handling

MetaDataGraphs.get_data — Method

get_data(g, label_s, label_d)

Retrieve edge-level metadata for the edge between the vertices with labels label_s and label_d.

MetaDataGraphs.get_data — Method

get_data(g, label)

Retrieve vertex-level metadata for the vertex with label label.

MetaDataGraphs.set_data! — Method

sett_data!(g, label_s, label_d, data)

Set edge-level metadata to value data for the edge between the vertices with labels label_s and label_d.

MetaDataGraphs.set_data! — Method

set_data!(g, label, data)

Set vertex-level metadata to value data for the vertex with label label.

Edge weights

Graphs.weights — Method

weights(g)

Compute the edge weights for an AbstractDataGraph.

If the edge data is of type ED = Nothing, return a lazy matrix full of ones.

If the edge data is of type ED <: Real, fill a sparse matrix with it.

If the edge data is of another type ED, try to use the attribute edge_data.weight for each edge.

Dict behavior

Base.getindex — Method

getindex(g, label)

Alias for get_vertex(g, label), intended for dictionary-like use.