AbstractDataGraph

Type

MetaDataGraphs.DataDiGraph — Type

DataDiGraph{T,VL,VD,ED,GD} <: AbstractDataGraph{T}

Structure for graphs with metadata based on adjacency list storage.

Fields

ne::Int: number of edges

fadjlist::Vector{Vector{Int}}: forward adjacency lists

badjlist::Vector{Vector{Int}}: backward adjacency lists

labels::Vector{VL}: list of vertex labels

vertices::Dict{VL,T}: dictionary mapping each vertex label to the associated integer index v

vertex_data::Vector{VD}: list of vertex data objects, indexed by integers v

edge_data::Vector{Vector{ED}}: list of edge data objects, indexed by s first and d_index second, where d_index is the rank of d among the outneighbors of s

graph_data::GD: single graph data object

MetaDataGraphs.DataDiGraph — Method

DataDiGraph{T}(VL, VD, ED, graph_data)

Constructor taking only label and data types to create an empty DataDiGraph.

Basic access

Base.haskey — Method

haskey(g, label)

Check whether a vertex with label label exists.

MetaDataGraphs.get_label — Method

get_label(g, v)

Retrieve the label associated with vertex v.

MetaDataGraphs.get_vertex — Method

get_vertex(g, label)

Retrieve the vertex associated with label `label.

Graphs.jl interface

Graphs.SimpleGraphs.add_edge! — Method

add_edge!(g::DataDiGraph, s, d, data)

Add an edge to g by specifying the vertices s and d along with the associated data.

Graphs.SimpleGraphs.add_vertex! — Method

add_vertex!(g::DataDiGraph, label, data)

Add a vertex to g by specifying its label along with the associated data.

Metadata handling

MetaDataGraphs.get_data — Method

get_data(g, s, d)

Retrieve edge-level metadata for edge (s, d).

MetaDataGraphs.get_data — Method

get_data(g, v)

Retrieve vertex-level metadata for vertex v.

MetaDataGraphs.get_data — Method

get_data(g)

Retrieve graph-level metadata.

MetaDataGraphs.set_data! — Method

set_data!(g, s, d, data)

Set edge-level metadata to value data for edge (s, d).

MetaDataGraphs.set_data! — Method

set_data!(g, v, data)

Set vertex-level metadata to value data for vertex v.