API reference

Index

DataGraphs.AbstractDataGraph

DataGraphs.ArrayDataDiGraph

DataGraphs.DictDataGraph

Docstrings

DataGraphs.AbstractDataGraph — Type

AbstractDataGraph{T,VL,VD,ED,GD} <: AbstractGraph{T}

General template for graphs with metadata. Here,

T<:Integer is the type of vertices

VL is the type of vertex labels (cannot be a subtype of Integer)

VD is the type of vertex data objects

ED is the type of edge data objects

GD is the type of the graph data object

DataGraphs.ArrayDataDiGraph — Type

ArrayDataDiGraph{T,VL,VD,ED,GD} <: AbstractDataGraph{T}

Structure for graphs with metadata based on adjacency list storage.

Fields

ne::Int: number of edges

fadjlist::Vector{Vector{Int}}: forward adjacency lists

badjlist::Vector{Vector{Int}}: backward adjacency lists

labels::Vector{VL}: list of vertex labels

vertices::Dict{VL,T}: dictionary mapping each vertex label to the associated integer index v

vertex_data::Vector{VD}: list of vertex data objects, indexed by integers v

edge_data::Vector{Vector{ED}}: list of edge data objects, indexed by s first and d_index second, where d_index is the rank of d among the outneighbors of s

graph_data::GD: single graph data object

DataGraphs.DictDataGraph — Type

DictDataGraph{T,G<:AbstractGraph{T},VL,VD,ED,GD} <: AbstractDataGraph{T}

Structure for graphs with metadata based on an underlying dataless graph.

Fields

graph::G: underlying graph, following the Graphs.jl AbstractGraph{T} interface

labels::Vector{VL}: list of vertex labels

vertices::Dict{VL,T}: dictionary mapping each vertex label to the associated integer index v

vertex_data::Vector{VD}: list of vertex data objects, indexed by integers v

edge_data::Dict{Tuple{T,T},ED}: list of edge data objects, indexed by integer tuples (s,d)

graph_data::GD: single graph data object